chore: Extract debug related code in monomorphization
And apply more code review feedback
